requirements.txt
torch
torchvision
gradio
scikit-learn
seaborn
matplotlib
tqdm
Pillow
